Very simple matching test

+++ cli/macrostrat/cli/subsystems/sgp/match/__init__.py
@@ -15,9 +15,7 @@ from sqlalchemy.dialects.postgresql import insert
 from sqlalchemy.sql import text
 
 from macrostrat.cli.database import get_db
-from macrostrat.core import app
-
-from ..utils import get_sgp_samples, stored_procedure, write_to_file
+from macrostrat.core.console import err_console as console
 from .clean_strat_name import (
     StratNameTextMatch,
     StratRank,
@@ -25,6 +23,7 @@ from .clean_strat_name import (
     clean_strat_name_text,
     format_name,
 )
+from ..utils import get_sgp_samples, stored_procedure, write_to_file
 
 here = Path(__file__).parent
 
@@ -62,12 +61,14 @@ class MatchComparison(Enum):
 
 def import_sgp_data(
     out_file: Path = None,
+    *,
     verbose: bool = False,
     sample: int = None,
     column: int = None,
     match: list[MatchType] = None,
     reset: bool = False,
     only_macrostrat: bool = False,
+    write: bool = False,
     # comparison: MatchComparison = MatchComparison.Included.value,
 ):
     """
@@ -79,7 +80,7 @@ def import_sgp_data(
     if only_macrostrat:
         samples = samples[samples["match_set"] != "Age from other sources"]
 
-    app.console.print(
+    console.print(
         f"Got {len(samples)} samples from the SGP database.",
         style="green",
     )
@@ -94,7 +95,7 @@ def import_sgp_data(
         columns=["member", "formation", "group", "verbatim_strat"], inplace=True
     )
 
-    app.console.print("\nMacrostrat columns", style="bold")
+    console.print("\nMacrostrat columns", style="bold")
 
     # Join samples to columns
     columns = get_columns_data_frame()
@@ -114,21 +115,21 @@ def import_sgp_data(
     n_matched = len(res[res["col_id"].notnull()])
     n_not_matched = len(res[res["col_id"].isnull()])
 
-    app.console.print(
+    console.print(
         f"Matched {n_matched} of {n_total} samples to a Macrostrat column.",
         style="green",
     )
 
     if n_too_many > 0:
-        app.console.print(
+        console.print(
             f"{n_too_many} samples have multiple matched columns.", style="yellow"
         )
     if n_not_matched > 0:
-        app.console.print(
+        console.print(
             f"{n_not_matched} samples have no matched columns.", style="yellow"
         )
     if n_too_many + n_not_matched == 0:
-        app.console.print("All measurements matched to a single column.", style="green")
+        console.print("All measurements matched to a single column.", style="green")
 
     # Augment the samples with the matched column
     samples = samples.join(res, on="sample_id", how="left")
@@ -150,24 +151,24 @@ def import_sgp_data(
         .apply(lambda x: tuple(sorted(x)))
     )
 
-    app.console.print("\nStratigraphic names", style="bold")
+    console.print("\nStratigraphic names", style="bold")
 
-    app.console.print(f"Found {len(counts)} unique stratigraphic name groups.")
+    console.print(f"Found {len(counts)} unique stratigraphic name groups.")
 
     if verbose:
         for ix, row in counts.iterrows():
-            app.console.print(f"{row['count']} samples in column {row['col_id']}")
-            app.console.print("  " + format_names(row["strat_names"]))
-        app.console.print()
+            console.print(f"{row['count']} samples in column {row['col_id']}")
+            console.print("  " + format_names(row["strat_names"]))
+        console.print()
 
     # Summarize the counts
     _match = counts["strat_names"].apply(lambda x: len(x) > 0)
 
-    print_counts(app.console, "candidate stratigraphic name", counts[_match], n_total)
+    print_counts(console, "candidate stratigraphic name", counts[_match], n_total)
 
     # Add empty columns for match information
 
-    app.console.print("\nUnit matching", style="bold")
+    console.print("\nUnit matching", style="bold")
 
     type_list = join_items(
         [f"[underline]{v.value}[/underline]" for v in get_match_types(match)],
@@ -175,7 +176,7 @@ def import_sgp_data(
         last=" or ",
     )
 
-    app.console.print(
+    console.print(
         "Matching based on",
         type_list,
         "\n",
@@ -212,8 +213,8 @@ def import_sgp_data(
     # Rename columns
 
     if matches.empty:
-        app.console.print("No matches found.")
-        return
+        console.print("No matches found.")
+        return None
 
     # Add interpreted age information by taking the midpoint ages
     ix = matches.columns.get_loc("max_age")
@@ -256,7 +257,8 @@ def import_sgp_data(
         # Convert field to be more suitable for export
 
         write_to_file(samples, out_file)
-    else:
+
+    if write:
         # Check if table exists
         if reset:
             M.engine.execute(text("TRUNCATE TABLE integrations.sgp_matches"))
@@ -275,8 +277,10 @@ def import_sgp_data(
             dtype={"geom": Geometry("POINT", srid=4326)},
         )
 
+    return samples
 
-def log_match_data(row, _match, *, verbose=False, console=app.console):
+
+def log_match_data(row, _match, *, verbose=False, console=console):
     message = "[red bold]no match[/]"
     if _match is not None:
         message = "[green bold]" + _match.strat_name_clean + "[/]"
@@ -304,7 +308,7 @@ def log_match_data(row, _match, *, verbose=False, console=app.console):
         console.print()
 
 
-def log_match_row(row, *, verbose=False, console=app.console):
+def log_match_row(row, *, verbose=False, console=console):
     match = None
     if not isna(row["unit_id"]):
         match = row
@@ -313,7 +317,7 @@ def log_match_row(row, *, verbose=False, console=app.console):
 
 def log_matches(verbose: bool = False):
     """Log SGP matches to the console"""
-    console = app.console
+    console = console
     M = get_db()
 
     # Get the matches
